requirements.txt
# Web Infrastructure
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests

# Compute Base (pinned for Python 3.10 wheel compatibility)
numpy==1.26.4
scikit-learn==1.3.2

# Intelligence & Agents
langgraph
groq
